CHORE: Refactoring
