+++ Frontend/src/App.jsx
@@ -1,8 +1,8 @@
 import { Routes, Route } from "react-router-dom";
 import LandingPage from "./pages/LandingPage";
 import SignIn from "./components/forms/SignIn";
+import SignUp from "./components/forms/SignUp";
 import Newsignup from "./components/forms/Newsignup";
-
 
 import TransactionSuccess from "./pages/TransactionSuccess";
 import PaymentSuccess from "./pages/PaymentSuccess";
@@ -11,32 +11,32 @@ import OtpPage from "./pages/OtpPage";
 import ProcessingPage from "./components/ProcessingPage";
 import PaymentModal from "./components/PaymentModal";
 
-// import FindMeds from "./pages/FindMeds";
 import SearchPage from "./pages/SearchPage";
 import SearchpageSummary from "./pages/SearchpageSummary";
-import UpdatedCart from "./pages/UpdatedCart";
+import OrderSummary from "./pages/OrderSummary";
 import { useState } from "react";
 import ProfileSignup from "./pages/ProfileSignup";
 import Dashboard from "./pages/Dashboard";
-
+import FindMedsLoading from "./components/FindMedsLoading";
+import PharamacySelection from "./pages/PharmacySelection";
 
 function App() {
   const [selectedMedicines, setSelectedMedicines] = useState([]);
   return (
     <Routes>
+      {/* Patients Routing */}
       <Route path="/" element={<LandingPage />} />
-      <Route path="/signup" element={<Newsignup />} />
+      <Route path="/signup-phone" element={<Newsignup />} />
+      <Route path="/signup-email" element={<SignUp />} />
       <Route path="/signin" element={<SignIn />} />
 
-      <Route path="/transactionsuccess" element={<TransactionSuccess />} />
-      <Route path="/paymentsuccess" element={<PaymentSuccess />} />
-      <Route path="/otppage" element={<OtpPage/>} />
+      <Route path="/transaction-success" element={<TransactionSuccess />} />
+      <Route path="/payment-success" element={<PaymentSuccess />} />
+      <Route path="/otp-page" element={<OtpPage />} />
 
       <Route path="/processing" element={<ProcessingPage />} />
 
-      <Route path="/paymentmodal" element={<PaymentModal />} />
-      <Route path="/dashboard" element={<Dashboard />} />
-
+      <Route path="/payment-modal" element={<PaymentModal />} />
 
       <Route
         path="/findmeds"
@@ -48,6 +48,9 @@ function App() {
         }
       />
 
+      <Route path="/findmeds-loading" element={<FindMedsLoading />} />
+
+      <Route path="/pharmacy-selection" element={<PharamacySelection />} />
 
       <Route
         path="/search-summary"
@@ -58,9 +61,13 @@ function App() {
           />
         }
       />
-      <Route path="/updatedcart" element={<UpdatedCart />} />
+
+      <Route path="/order-summary" element={<OrderSummary />} />
 
       <Route path="/complete-profile" element={<ProfileSignup />} />
+
+      {/* Pharmacy Routing */}
+      <Route path="/dashboard" element={<Dashboard />} />
     </Routes>
   );
 }
