+++ Frontend/src/components/FindMedsLoading.jsx
@@ -1,14 +1,20 @@
 import React from 'react'
+import "../styles/ProcessingPage.css";
 
 const FindMedsLoading = () => {
   return (
-    <div className='loading-container'>
-<div className='loading-spinner'></div>
-    <h2>
-        Finding Meds......
-    </h2>
+    <div className="container">
+      <div className="card">
+        <div className="spinner"></div>
+        <p>Finding Pharmacies, please wait.</p>
+      </div>
     </div>
   )
 }
 
 export default FindMedsLoading
+
+
+
+
+
