+++ Frontend/src/pages/PharmacySelection.jsx
